:sparkles: feat: feignconfig 적용
