+++ payment/src/main/java/com/quit/payment/infrastructure/configuration/PaymentFeignConfig.java
@@ -1,16 +1,9 @@
 package com.quit.payment.infrastructure.configuration;
 
-import feign.Logger;
-import feign.Request;
 import feign.RequestInterceptor;
-import feign.Retryer;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 
-import java.util.concurrent.TimeUnit;
-
-@Configuration
 @RequiredArgsConstructor
 public class PaymentFeignConfig {
 
@@ -19,21 +12,6 @@ public class PaymentFeignConfig {
     @Bean
     public RequestInterceptor paymentAuthInterceptor() {
         return new PaymentAuthInterceptor(paymentProperties);
-    }
-
-    @Bean
-    public Request.Options requestOptions() {
-        return new Request.Options(3, TimeUnit.SECONDS, 30, TimeUnit.SECONDS, true);
-    }
-
-    @Bean
-    public Retryer retryer() {
-        return new Retryer.Default(2000, 5000, 3);
-    }
-
-    @Bean
-    Logger.Level feignLoggerLevel() {
-        return Logger.Level.FULL;
     }
 
 }
